+++ crates/objc2/CHANGELOG.md
@@ -20,6 +20,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Implemented more `fmt` traits on `Retained`.
 * Implemented `Extend` trait on `Retained`.
 * Implemented `AsRef` in a forwarding fashion on `Retained`.
+* Implemented `PartialEq` and `PartialOrd` on `Retained` in a slightly more
+  generic way.
 
 ### Changed
 * **BREAKING**: Changed how you specify a class to only be available on the